+++ lib/casbin-ruby/config/config.rb
