+++ lib/casbin-ruby/effect/effector.rb
