+++ lib/casbin-ruby/effect/priority_effector.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'casbin/ruby/effect/effector'
+require 'casbin-ruby/effect/effector'
 
 module Casbin
   module Effect
