+++ lib/casbin-ruby/model/assertion.rb
