+++ lib/casbin-ruby/model/function_map.rb
