+++ lib/casbin-ruby/persist/adapter.rb
