+++ lib/casbin-ruby/rbac/role_manager.rb
