+++ lib/casbin-ruby/util.rb
