+++ lib/casbin-ruby/util/evaluator.rb
