+++ lib/casbin-ruby/util/thread_lock.rb
